+++ changelog.md
@@ -1,13 +1,15 @@
 # Changelog
 
 ## Alpha 2.0
-- 
+- Allow users to use the app without logging in
+- Fix Android Studio warning messages
+- Fixed minor bugs
 
 ## Alpha 1.0
 - Signup and login using Firebase Auth
 - Create and store palettes in Firebase Realtime DB
   - A palette consist of 2 X 3 blocks
   - Store .png of the block palette in Firebase Storage
-- Search palettes using blocks 
+- Search palettes using blocks
 - Like palettes using heart button
 - Share .png of palette using share button